Add files via upload
I have used a sample dataset which I downloaded from Kaggle.com to do this dashboard. The dashboard demonstrates "The Winter Olympics for a period between 1924 and 2014". From the analysis, we understand that the Speedskating Sport won the most medals, followed by Cross-Country Skiing - which also means it's one of the oldest Olympic sport. Norway had bagged the highest number of medals and the position changes as we choose different year from the line graph - example: In 1948, Switzerland went home with most medals. Also, we can see the that the total number of countries participating in Winter Olympics increases steadily through 2014 by clicking on a 'year' in line graph. Another observation is that, men had won most medals so far compared to women and mixed age group. Players between age group of 23 to 28 had taken most medals.