Add files via upload
I have worked on a Sample Dataset. The ABC Cookie Company Sales Report showcases the companies sales for the year 2019. We could explicitly see the companies Revenue, Units sold, Profit and there Top 3 products sold in the market. We can see that France is the country with highest revenue followed by United States in second and Canada in the third position. With interactive tiles, on clicking the country we can understand the profit by product and units sold for that country. Also on product level Chocolate chip is the most sold and Fortune cookie is the least sold across all countries. Maybe as a suggestion from the insights we can say that by changing the packaging or giving offers the sales of fortune cookies can be improved.